Remove unused variables
Just a bit of cleanup. `Tracker::sessionLog($uuid)` was called 2 times here. The variable 'log' passed to the view is also unused because it just makes an ajax request with the uuid.

Please note that this could be a breaking change if someone extended the log view and used the `$log` variable without extending the controller.